[math]: add type safety to `include/math/` operations
I've been holding off improving typing and type ergonomics
in the kernel for a while, but I figured that I should
probably stop beating around the bush at some point.

A bunch of CMakelists.txt compiler warn flags regarding
types were added, and I added many type safety/`typecheck`
macros to `compiler.h`, and began using them in `include/math`.

Many files around the kernel, particularly drivers and
small things were touched to use the more ergonomic macros.

However, TODO: we will probably need to begin using the various BIT_ macros
and whatnot inside of core `enum x_flags` definitions (like thread_flags),
but I'm not lumping that into this commit because it's best changed
in its own batch of changes, with its own commit so as to keep the
blame of those files relatively ok.

Tags: adds 1 TODO

+++ include/dbg.h
@@ -3,7 +3,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#define STACK_TRACE_MAX_DEPTH 64
+#define STACK_TRACE_MAX_DEPTH ((size_t) 64)
 
 void debug_print_registers();
